+++ integration/nwo/token/common/ppmgen.go
@@ -11,14 +11,13 @@ import (
 	"path/filepath"
 	"strconv"
 
-	msp2 "github.com/hyperledger-labs/fabric-token-sdk/token/core/identity/msp"
-
 	msp "github.com/IBM/idemix"
 	math3 "github.com/IBM/mathlib"
 	"github.com/hyperledger-labs/fabric-smart-client/platform/fabric/core/generic/msp/x509"
 	"github.com/hyperledger-labs/fabric-token-sdk/integration/nwo/token/generators"
 	"github.com/hyperledger-labs/fabric-token-sdk/integration/nwo/token/topology"
 	"github.com/hyperledger-labs/fabric-token-sdk/token/core/fabtoken"
+	msp2 "github.com/hyperledger-labs/fabric-token-sdk/token/core/identity/msp"
 	cryptodlog "github.com/hyperledger-labs/fabric-token-sdk/token/core/zkatdlog/crypto"
 	"github.com/pkg/errors"
 )
